# docs/api-guides/sqlrs-engine.openapi.yaml
openapi: 3.1.0
info:
  title: The sqlrs engine API
  version: 0.1.0
  description: |
    Local sqlrs engine HTTP API (MVP).
    This file is the design source of truth for the engine API.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
security:
  - bearerAuth: []
servers:
  - url: http://127.0.0.1:{port}
    description: Local engine endpoint (from engine.json)
    variables:
      port:
        default: "8080"
paths:
...
  /v1/prepare-jobs:
    get:
      operationId: listPrepareJobs
      summary: List prepare jobs
      description: Returns prepare jobs.
      tags:
        - prepare
      parameters:
        - in: query
          name: job
          schema:
            type: string
          description: Filter by job id prefix.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PrepareJobEntry"
            application/x-ndjson:
              schema:
                type: string
                description: Newline-delimited JSON stream of PrepareJobEntry objects.
        "400":
          description: Invalid filter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
    post:
      operationId: createPrepareJob
      summary: Start a prepare job
      description: |
        Starts a prepare job and returns a job reference.
        The engine enforces deterministic defaults for `psql` (`-X` and
        `-v ON_ERROR_STOP=1`) and rejects connection flags.
        When `plan_only` is true, the job computes a plan and does not create
        an instance; the plan tasks are returned in the job status.
      tags:
        - prepare
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrepareJobRequest"
      responses:
        "201":
          description: Created
          headers:
            Location:
              schema:
                type: string
              description: Canonical job status URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrepareJobAccepted"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/tasks:
    get:
      operationId: listTasks
      summary: List tasks
      description: Returns task queue entries across jobs.
      tags:
        - tasks
      parameters:
        - in: query
          name: job
          schema:
            type: string
          description: Filter by job id (exact match).
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskEntry"
            application/x-ndjson:
              schema:
                type: string
                description: Newline-delimited JSON stream of TaskEntry objects.
        "400":
          description: Invalid filter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
  /v1/config:
    get:
      operationId: getConfig
      summary: Get server config
      description: |
        Returns the server configuration.
        When `path` is provided, returns the value at that path.
        When `effective` is true, defaults are applied before returning the value.
      tags:
        - config
      parameters:
        - in: query
          name: path
          required: false
          schema:
            type: string
          description: JS-style path (e.g. `orchestrator.jobs.maxIdentical`).
        - in: query
          name: effective
          required: false
          schema:
            type: boolean
          description: Return value after applying defaults.
      responses:
        "200":
          description: Config value or full config object.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ConfigValue"
                  - $ref: "#/components/schemas/ConfigMap"
        "400":
          description: Invalid path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
        "404":
          description: Path not found
    patch:
      operationId: setConfig
      summary: Set config value
      description: |
        Sets a configuration value by path.
        The value is validated against the server schema and semantic rules.
      tags:
        - config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigSetRequest"
      responses:
        "200":
          description: Updated config value.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigValue"
        "400":
          description: Invalid value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
    delete:
      operationId: removeConfig
      summary: Remove config value
      description: |
        Removes a configuration value by path.
      tags:
        - config
      parameters:
        - in: query
          name: path
          required: true
          schema:
            type: string
          description: JS-style path (e.g. `orchestrator.jobs.maxIdentical`).
        - in: query
          name: effective
          required: false
          schema:
            type: boolean
          description: Return value after applying defaults.
      responses:
        "200":
          description: Updated config value.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigValue"
        "400":
          description: Invalid path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
        "404":
          description: Path not found
  /v1/config/schema:
    get:
      operationId: getConfigSchema
      summary: Get config schema
      description: Returns the JSON Schema for server config validation.
      tags:
        - config
      responses:
        "200":
          description: JSON Schema
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigSchema"
        "401":
          description: Unauthorized
  /v1/names:
    get:
      operationId: listNames
      summary: List names
      description: Returns name bindings.
      tags:
        - names
      parameters:
        - in: query
          name: instance
          schema:
            type: string
          description: Filter by instance id.
        - in: query
          name: state
          schema:
            type: string
          description: Filter by state id.
        - in: query
          name: image
          schema:
            type: string
          description: Filter by base image id.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NameEntry"
            application/x-ndjson:
              schema:
                type: string
                description: Newline-delimited JSON stream of NameEntry objects.
        "401":
          description: Unauthorized
  /v1/names/{name}:
    get:
      operationId: getName
      summary: Get a name binding
      description: Returns a single name binding.
      tags:
        - names
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NameEntry"
        "404":
          description: Not found
        "401":
          description: Unauthorized
  /v1/instances:
    get:
      operationId: listInstances
      summary: List instances
      description: Returns instances.
      tags:
        - instances
      parameters:
        - in: query
          name: id_prefix
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{8,}$"
          description: Case-insensitive hex id prefix (min 8 chars).
        - in: query
          name: state
          schema:
            type: string
          description: Filter by state id.
        - in: query
          name: image
          schema:
            type: string
          description: Filter by base image id.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InstanceEntry"
            application/x-ndjson:
              schema:
                type: string
                description: Newline-delimited JSON stream of InstanceEntry objects.
        "400":
          description: Invalid filter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
  /v1/instances/{instanceId}:
    get:
      operationId: getInstance
      summary: Get an instance
      description: |
        Returns a single instance by id. If the path segment does not match the
        instance id format, it is treated as a name alias. If it matches the id
        format, the engine first attempts id lookup, then falls back to name.
        When resolved by name, the response is a temporary redirect to the
        canonical id-based URL.
      tags:
        - instances
      parameters:
        - in: path
          name: instanceId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstanceEntry"
        "307":
          description: Temporary redirect to canonical instance id URL
          headers:
            Location:
              schema:
                type: string
              description: Canonical instance URL.
        "404":
          description: Not found
        "401":
          description: Unauthorized
    delete:
      operationId: deleteInstance
      summary: Delete an instance
      description: |
        Deletes a single instance by id. This operation is idempotent.
        The response includes a deletion tree with blocked reasons when removal
        is not possible.
      tags:
        - instances
      parameters:
        - in: path
          name: instanceId
          required: true
          schema:
            type: string
        - in: query
          name: force
          schema:
            type: boolean
          description: Ignore active connections.
        - in: query
          name: dry_run
          schema:
            type: boolean
          description: |
            Return the deletion tree without making changes.
            When true, the response is always 200 and `outcome` indicates whether
            deletion would succeed.
      responses:
        "200":
          description: OK (deleted or dry-run result)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Blocked by safety rules (dry_run=false only)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
  /v1/runs:
    post:
      operationId: runCommand
      summary: Run a command against an instance
      description: |
        Executes a run kind command against an existing instance and streams
        stdout/stderr/exit events. The stream may also include log events for
        instance recovery when a missing container is recreated. If `command`
        is omitted, the default command for the selected kind is used.
      tags:
        - run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunRequest"
      responses:
        "200":
          description: Streamed run output
          content:
            application/x-ndjson:
              schema:
                $ref: "#/components/schemas/RunEvent"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
        "404":
          description: Instance not found or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflicting run arguments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  schemas:
...
    PrepareJobRequest:
      oneOf:
        - $ref: "#/components/schemas/PrepareJobRequestPsql"
      discriminator:
        propertyName: prepare_kind
        mapping:
          psql: "#/components/schemas/PrepareJobRequestPsql"
...
    ConfigValue:
      type: object
      additionalProperties: false
      required:
        - path
        - value
      properties:
        path:
          type: string
        value:
          $ref: "#/components/schemas/JSONValue"
    ConfigMap:
      type: object
      additionalProperties: true
      description: Full server config object.
...
    PrepareJobEntry:
      type: object
      additionalProperties: false
      required:
        - job_id
        - status
        - prepare_kind
        - image_id
      properties:
        job_id:
          type: string
        status:
          type: string
          enum: [queued, running, succeeded, failed]
        prepare_kind:
          type: string
        image_id:
          type: string
        plan_only:
          type: boolean
          description: True when the job computes a plan only.
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: string
...